+++ activerecord/test/models/book_destroy_async.rb
@@ -22,3 +22,21 @@ class BookDestroyAsyncWithScopedTags < ActiveRecord::Base
   has_many :taggings, as: :taggable, class_name: "Tagging"
   has_many :tags, -> { where name: "Der be rum" }, through: :taggings, dependent: :destroy_async
 end
+
+class BookDestroyAsyncWithDestroyingAsyncCallbackEssays < ActiveRecord::Base
+  self.table_name = "books"
+
+  has_many :essays,
+    dependent: :destroy_async,
+    class_name: "EssayWithDestroyingAsyncCallback",
+    foreign_key: "book_id"
+end
+
+class BookDestroySyncWithDestroyingAsyncCallbackEssays < ActiveRecord::Base
+  self.table_name = "books"
+
+  has_many :essays,
+    dependent: :destroy,
+    class_name: "EssayWithDestroyingAsyncCallback",
+    foreign_key: "book_id"
+end