+++ activerecord/test/models/essay_destroy_async.rb
@@ -10,3 +10,16 @@ end
 
 class ShortEssayDestroyAsync < EssayDestroyAsync
 end
+
+class EssayWithDestroyingAsyncCallback < ActiveRecord::Base
+  self.table_name = "essays"
+  around_destroy :callback, if: :destroying_async?
+
+  private
+    def callback
+      before_destroying_async
+      yield
+    end
+
+    def before_destroying_async; end
+end